[PDFJS] Loading dynamically instead of build time
